+++ core-ng/src/main/java/core/framework/http/HTTPResponse.java
@@ -1,7 +1,5 @@
 package core.framework.http;
 
-import core.framework.api.http.HTTPStatus;
-
 import java.nio.charset.Charset;
 import java.util.Map;
 
@@ -11,13 +9,13 @@ import static java.nio.charset.StandardCharsets.UTF_8;
  * @author neo
  */
 public final class HTTPResponse {
-    public final HTTPStatus status;
+    public final int statusCode;
     public final byte[] body;
     public final Map<String, String> headers;   // headers key is case insensitive
     public final ContentType contentType;
 
-    public HTTPResponse(HTTPStatus status, Map<String, String> headers, byte[] body) {
-        this.status = status;
+    public HTTPResponse(int statusCode, Map<String, String> headers, byte[] body) {
+        this.statusCode = statusCode;
         this.headers = headers;
         this.body = body;