+++ core-ng/src/test/java/core/framework/http/HTTPResponseTest.java
@@ -1,6 +1,5 @@
 package core.framework.http;
 
-import core.framework.api.http.HTTPStatus;
 import core.framework.util.Strings;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class HTTPResponseTest {
     @Test
     void text() {
-        var response = new HTTPResponse(HTTPStatus.OK, Map.of(HTTPHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.toString()), Strings.bytes("value"));
+        var response = new HTTPResponse(200, Map.of(HTTPHeaders.CONTENT_TYPE, ContentType.TEXT_PLAIN.toString()), Strings.bytes("value"));
 
         assertThat(response.contentType.toString()).isEqualTo(ContentType.TEXT_PLAIN.toString());
         assertThat(response.text()).isEqualTo("value");